WIP: Wrapper trait
XRB can't implement any cornflakes traits on Option, because it defines neither.  Therefore, it needs a way to be able to do that without explicitly being the one to implement it on Option... with this Wrapper trait, XRB can implement it on T, and then cornflakes can implement for Option<T> appropriately.

Note that the more specific implementations for DataSize where T: DataSize + StaticDataSize were removed because they were causing errors.  That's not because they shouldn't be there - if the errors can be fixed, they should stay.